+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+* Export MUI `Box`
+* Export our three themes and three palettes
+
 ## 2.7.1 (2019-07-15)
 
 * Add MessageTile (and three subcomponents) for upcoming promos (https://github.com/conversation/ui/pull/82)